measurement: count measurements for FACCH/H twice.
The FACCH/H replaces to voice frames in TCH/H. This means during the
transmission of a FACCH one measurement value less is handed up to the
higher layers. This needs to be compensated by adding a the FAACH/H
measurement twice (as if it were from two voice frames).

Change-Id: Id533bb3207bac40d00d77b484619f66ea5c0774d
Related: OS#4799